Split script into multiple files
